requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
numpy>=1.24.0
